Improve bench harness measurement integrity (#58)
## Summary

Make the benchmark harness fail closed so performance reports reflect the workflow that was actually requested and completed.

The previous tooling could accept incomplete captures, combine unrelated runs, misclassify sleep and storage events, or apply budgets to the wrong samples. This made some results, especially page-turn latency, wake timing, and cold versus warm storage measurements, less reliable than the reports suggested.

## Changes

- Add structured run metadata, completion records, boot boundaries, and per-run workflow validation.
- Pair page-turn inputs with the correct Reading render and stop after the requested number of completed pairs.
- Validate sleep completion, wake ordering, failed sleep phases, and requested cycle counts.
- Separate RAM hits, warm opens, cold builds, background builds, and catalog outcomes.
- Verify requested storage modes using explicit successful evidence.
- Scope coverage checks, signal validation, pooled statistics, and budgets to the correct run and workflow.
- Reject incomplete captures, malformed telemetry, unknown result values, and invalid budget files in strict mode.
- Make benchmark logging optional in firmware while retaining compile-time format checking.
- Standardize the Python and Ruff versions used locally and in CI.
- Expand the benchmark tests and documentation to cover these contracts.

## Result

Strict mode now verifies that each capture completed its requested workflow, produced the required telemetry, and was evaluated against a valid applicable budget.

The canonical X3 benchmark measurements should be collected again from the merged `main` commit because the measurement contract has changed substantially.

+++ tools/fontgen_common.py
@@ -1,17 +1,15 @@
-from pathlib import Path
 import os
 import struct
+from pathlib import Path
 
 from PIL import Image, ImageDraw
-
 
 ADVANCE_SCALE = 16
 DEFAULT_THRESHOLD = 128
 MIN_KERNING_ADJUST_FP = 8
 MAX_KERNING_ENTRIES = 1024
 KERNING_CODEPOINTS = frozenset(
-    list(range(0x20, 0x7F))
-    + [0x2018, 0x2019, 0x201C, 0x201D, 0x2013, 0x2014, 0x2026]
+    list(range(0x20, 0x7F)) + [0x2018, 0x2019, 0x201C, 0x201D, 0x2013, 0x2014, 0x2026]
 )
 
 
@@ -26,7 +24,7 @@ THRESHOLD = text_render_threshold()
 
 
 def advance_fp(font, text: str) -> int:
-    return max(int(round(font.getlength(text) * ADVANCE_SCALE)), ADVANCE_SCALE)
+    return max(round(font.getlength(text) * ADVANCE_SCALE), ADVANCE_SCALE)
 
 
 def rasterize_glyph(font, code: int):
@@ -337,7 +335,7 @@ def kerning_entries(font_path: Path, cps, px: int):
 
     entries = []
     for (left_glyph, right_glyph), value in adjustments.items():
-        adjust_fp = int(round(value * px * ADVANCE_SCALE / units_per_em))
+        adjust_fp = round(value * px * ADVANCE_SCALE / units_per_em)
         if abs(adjust_fp) < MIN_KERNING_ADJUST_FP:
             continue
         for left in cps_by_glyph.get(left_glyph, []):
